statespec: runtime context type-checking (validate_context)
apply() now checks that the runtime values a service supplies match the
declared field-type tags (not just that fields are present) — so a snapshot
with "0" (str) where an int was declared fails loud rather than silently
flipping a comparison. age_days is tagged `decimal` (the service computes a
float). Closes the "context trusted but not type-validated" gap from review.

+++ backend/app/statespec/submission_spec.py
@@ -41,8 +41,9 @@ SUBMISSION_SPEC = StateSpec(
     name="submission",
     title="Submission moderation lifecycle",
     states=_STATES,
-    # Context contract — what a service snapshot must provide.
-    fields={"status": "str", "age_days": "int"},
+    # Context contract — what a service snapshot must provide. age_days is the
+    # days since created_at (the service computes it as a float).
+    fields={"status": "str", "age_days": "decimal"},
     initial="pending",
     terminal=frozenset({"approved", "rejected", "expired"}),
     transitions=(
